requirements.txt
babel==2.10.1
python-dateutil==2.6.0
flask-moment==0.11.0
flask-wtf==1.0.1
flask_sqlalchemy==2.4.4
